requirements.txt
flask
wheel
gunicorn
rsa
requests
google-cloud-language
google-cloud-secret-manager